Bugfixes and first interpreter speed test
The interpreter ran at about 24m rewrites per second on my first test,
on Apple's M1. That's about 6x slower than usual speeds on the version
compiled to C. The interpreter will probably be used by Kindelia. 24m
isn't enough, it should be at least around 60m. But it is written in a
very convoluted way, so it might not be hard to optimize it. Next steps
are to optimize the interpreter and finish porting the C compiler