Two small changes in 'supperposed application'
I propose two changes in the Readme because these are subtly unexpected expressions that may be significant but go unexplained. So it is either a typo, or something more advanced that is not explained.

`({λx(x) λy(x)} 10)` -> `({λx(x) λy(y)} 10)` (There does not seem to be any use of the shared variable `x`, and variables cannot be shared anyway).`

The other one seems wrong, as the notation `{a b} = c` has never been introduced. Even if at this point it may make sense to replace `dup a b = c` by `{a b} = c`, a note about it would be welcome.

```
({a b} c)
----------------------- App-Sup
{x0 x1} = c --> dup x0 x1
{(a x0) (b x1)}
```