Dynamic dup labels: less restrictive cloning
Global rewrite rules now generate a fresh dup label every time they are
called. This greatly increases the class of programs that don't exhaust
the cloning limitation. There is one problem, though: the label of a dup
node is stored as a 24-bit value inside its pointer. That means there
are only 2^24 possible labels. As such, label collisions, even though
rare, may happen, resulting in incorrect results. Regardless, this is
always more permissible than the old version, but it may cause result in
programs that work for small inputs, but fail for big enough inputs.
Note these programs would just fail for any input before, but the fact
they will work for small inputs, but have a small chance of failing for
large inputs, isn't ideal. The only solution I can think of is to create
a 128-bit version of HVM, which should be used for large inputs. Such
version would be able to store 64-bit labels, avoiding this issue.