Raise deadlock warning
HVM will deadlock on the following input:

  (MapReduce red map (Leaf x))   = (map x)
  (MapReduce red map (Node a b)) = (red (MapReduce red map a) (MapReduce red map b))

  (Gen 0) = (Leaf 1)
  (Gen n) = (Node (Gen (- n 1)) (Gen (- n 1)))

  (Main n) = (MapReduce λaλb(+ a b) λx(* x 2) (Gen 7))

This is caused because the combination of the task stealer with dup-node
locks can cause a thread to lock a dup node that another thread - which
is currently waiting for this thread - ends up waiting for too. I do not
currently have the time to investigate this issue, so HVM will just add
a deadlock detector and raise a warning in case a deadlock occurs. Given
my limited time and resources, for now, that's the best I can do.