include .root on Defs, allow node on index 0
rationale: currently, the root of GNet is accessible via Var(1), so, we
don't need the index 0 of the nodes buffer for that. also, since Var(0)
stands for "empty", we can have Con(0), allowing us to address node 0.

note that the allocator still skips node 0; it must be updated to allow
allocating it.