bugfixes and completed new num parser, see notes
HVM2-32 numbers are 29-bit values:
- 1-bit flip tag (used internally, by oper interaction)
- 4-bit numb typ (can be SYM, U24, I24 or F24)
- 24-bit value (the number itself)

The parser now allows parsing all these variants, as follows:

- SYM: parsed as `[+]`, `[-]`, `[*]`, etc. (see code)

- U24: parsed as a numeric decimal (1234) or hex (0x12AF)

- I24: same as U24, but always prefixed by a sign (+ or -)

- F24: parsed as a decimal float (3.14) with an optional sign

Moreover, the flip bit can be parsed with a `:` before the number.
This can be useful to write reduced OPR. For example, instead of:

    // λa λb (- a b)
    @sub = (a (b r))
      & [-] ~ $(a $(b r))

One can write:

    // λa λb (- a b)
    @sub = ($(:[-] $(b r)) (b r))

Which has the same effect, but with the redex consumed. The ':' is
necessary to flip the operation, otherwise it would become '(- b a)'.

The stringifier will now also correctly prettify the 29-bit value.