Prevent leaking local nodes
Previously, when a local node was linked to a global var, it would leak.
This demanded lots of checks all around the code to deal with cases
where a port we're reading is unavailable (not written yet). This change
delays leaking nodes to the same moment the page is saved to the global
buffer, making all these checks anymore.

Note that this is only possible because the kernel stops right after the
page is saved, and the leaked links are saved to the redex bag to be
executed on the next kernel launch. That means that all nodes that these
redexes refer to will then be available to all blocks. To perform the
hot page swapping optimization, we would like to, once again, handle
this case. But, instead of handling with a failure (as we did before),
we will actually be able to just loop and wait for the value to be
non-zero. That's because leaked nodes are saved at the same time the
page is uploaded, so it is guaranteed to be available soon. Previously,
that was not true, because a node could leak in the middle of the
block's execution, so, waiting for it could result in deadlocks.