Re-add no-leak policy, viable by an optimization
Now, when doing:

    link global_var <- local_var

Instead of delaying that "leak", we will just swap the link, as that
won't be a leak. This means we can't leak vars directly, only nodes; but
we can still leak vars indirectly via nodes. By doing so, though, we
manage to get the same performance back in cases that had degraded
performance. Moreover, since no-leak policy cleans up the code
considerably, there seems to be enough justification to considerit for
now. It would be very handy if we could avoid the delay buffer though,
as delaying links has consequences for performance.

With 0-leak policy, I believe page swapping might be approachable right
now. Of course, it will cause leaks if we perform the links when we
upload the page; but, since that means the page will be available, we
can handle the leaks by just busy-waiting on 0 values. I'm not so sure
that is a great idea, though, as that carries some assumptions on CUDA's
global memory behavior, which I'm not sure hold. Perhaps a better idea
is to just let the kernel exit and make page saving and loading as
lightweight as possible, and perhaps reduce the interaction count per
launch. I'm not sure how large is the delay between two kernel launches
though. Increasing the block count might also be a good approach to hide
the latency/slowdown caused by hanging blocks - if that's even the
issue. Profiling is necessary to decide how to proceed.