small polishments, consistent link order
I'm still not sure which order is the best. On my experimentation, a
demo tail-recursive program with one accumulating argument:

    @loop = (?(((a a) @loop$C0) (b c)) (b c))

    @loop$C0 = (a (b e))
      & @loop ~ (a (d e))
      & $(c d) ~ [*0000001]
      & $(b c) ~ [+0000001]

    @main = a
      & @loop ~ (20 (0 a))

uses little constant memory in all buffers. sadly, the sum of tree case:

    @Leaf = (a ((a b) (* b)))

    @Node = (a (b (* ((a (b c)) c))))

    @gen = (?((@gen$C0 @gen$C1) a) a)

    @gen$C0 = (a b)
      & @Leaf ~ (a b)

    @gen$C1 = ({a d} ({$([*0000002] $([+0000001] b)) $([*0000002] e)} g))
      & @Node ~ (c (f g))
      & @gen  ~ (a (b c))
      & @gen  ~ (d (e f))

    @sum = (((a a) (@sum$C0 b)) b)

    @sum$C0 = (a (b d))
      & @sum ~ (a $(:[+] $(c d)))
      & @sum ~ (b c)

    @main = b
      & @sum ~ (a b)
      & @gen ~ (8 (0 a))

seems to still need 50+ vars and nodes, even with small depths, which is
far above the current 32 limit of HVM-CUDA. I wonder if it would be
possible to evaluate programs like this one with less space.