granular leak - WIP
this commit removes the paging functionality, replacing it by a
single-port global memory allocator, and a granular leak function that
uploads any port from local to global memory. this is already working
for stress_test example, and we do reach the ~50b RPS on it even without
paging, which is promising - but "Sort 16 (radix)" is causing segfauls.
I don't know what is the reason, but most likely related to the leak
function, which ended up needing recursion and is very ugly. being able
to merge this commit would possibly be a good move, since it would allow
threads to operate more independently, as there would be no need for all
threads to abort because the page is full. it also allows us to pass the
32 node allocation limit; when the shared memory is full, we just alloc
from global memory, and everything works. in fact, even with no shared
memory at all, the stress_test can reach almost 20b. the 2.5x we get
from shard memory is essential, though. I'll keep trying to make this
work and figure out the cause of the segfault.