Granular leak instead of pages - done
This replaces the old paging system by a granular leak system where
nodes and vars can be moved from shared to global memory via a leak
interaction, that is part of the link interaction, and is triggered
when a global var is linked to a local node: `X <- (a b)`.

One advantage of this change is that it allows leaked links to be
processed immediately, rather than delayed to the next block. But the
main advantage is that it allows us to replace the old page system by a
more granular global allocator. This, in turn, allows us to implement a
"global allocation" mode, which blocks will fallback to when the shared
memory isn't enough for them to work. This way, programs that fill the
shared memory quickly can stay working for longer in a kernel launch.

Note that, with that, there are a few different choices regarding how
the global allocator works. We can alloc in such a way that neighbor
threads get neighbor locations of memory, as in:

    u32 lc = (GID() + (*g_put * TPG)) % G_NODE_LEN;

The advantage of doing it this way is that programs that are aligned and
allocating in sync will result in blocks that request continuous slices
of memory (like stress_test), which will be fast. Sadly, programs that
aren't allocating in such consistent pattern (like radix_sort) will not
benefit from that. The other option is to let each thread have a region
of memory which it allocates too:

    u32 lc = GID()*(G_NODE_LEN/TPG) + (*g_put%(G_NODE_LEN/TPG));

The advantage of doing it this way is that, when a single thread
allocates consecutive indices, they'll be cached. This makes radix_sort
2x faster, but makes stress_test 2x slower. Since stress_test uses
mostly shared memory, though, this is irrelevant to it, so, I decided to
go with the later version, that improves radix_sort and other more
chaotic programs. Sadly, this choice will mean that well-behaved,
regular programs (like shaders etc.) that do not manage to use only the
shared memory will be 2x slower than they could be.

There is a third approach, that is the best of both worlds. We can do a
__syncthreads() on each loop of the global allocator, so that threads
always allocate consecutive indices; i.e., keeping the 'g_put' of all
threads aligned, as they now increment at the same time. The problem of
doing that is that it requires modifying the interaction kernel in a way
that affects the maximum peak performance of the stress_test. In theory,
we could still have the best of all worlds, but I think it would require
having two versions of the evaluator: one optimized for shared memory,
and other optimized for global memory.

To keep the code simple, I'll leave it as is, but there is a lot of room
for improvement and optimization in multiple axis like these.