sanity checks for unfilled values
while this rarely happen, this can obviously happen with the granular
leaking version, so I'm adding these checks. on the enter() case, we can
just break, as it is just an optimization. on the link case, we need to
postpone the link. I only check when A is global, there is no need
otherwise (and it would impact performance). note that this can still in
theory happen on node_take and vars_take, but it has never happened in
practice, so I'm not adding the check yet just for now, to avoid
degrading performance (perhaps it can not actually happen, if CUDA
ensures the write order is consistent in certain parts of the code).
I'll revisit this later