improved, lightweight redex sharing in C
previously, the C version used a sharing strategy similar to CUDA. it
performed ~2^22 reductions in isolation, and then synchronized to share
redexes. this was good for cases like stress_test, where you share
redexes initially, and then spend a lot of time working alone. but in
cases like radix_sort, where sharing occurs more frequently, it
underperformed. and in the full sequential loop, it basically came to a
complete halt. here are the benchmarks:

- stress_test: 1200 MIPS
- radix_sort: 800 MIPS
- loop: 0.05 MIPS

now, instead, we'll just keep a global redex bag, and do task stealing.
this is much more flexible, and all cases will perform well:

- stress_test: 900 to 1200 MIPS (varies a lot)
- radix_sort: 1150 MIPS (stable)
- loop: 107 MIPS (stable)